Move Monte Carlo demo to standalone examples crate
