+++ examples/montecarlo-gain-target/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "montecarlo-gain-target"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+montycarlo = "0.1.0"
+rand = "0.8"
